Refactor page node utils
