+++ src/Plugins/Keymap.ts
@@ -32,7 +32,7 @@ import {
 } from "../utils/paragraph";
 import { isNodeEmpty } from "@tiptap/core";
 import { appendAndReplaceNode, deleteNode } from "../utils/nodes/node";
-import { getPageNodeByPageNum, isPageNode } from "../utils/page";
+import { getPageNodeByPageNum, isPageNode } from "../utils/nodes/page";
 
 const KeymapPlugin = keymap({
     ArrowLeft: (state, dispatch) => {
