+++ src/index.ts
@@ -29,7 +29,7 @@ import HeaderFooterNode from "./Nodes/HeaderFooter";
 import BodyNode from "./Nodes/Body";
 
 // === Utils ===
-export { isPageNode } from "./utils/page";
+export { isPageNode } from "./utils/nodes/page";
 export { getPageNumber, getThisPageNodePosition } from "./utils/pagination";
 export { getPageNumPaperSize, getPageNodePaperSize } from "./utils/paperSize";
 export { getPageNumPaperColour, getPageNodePaperColour } from "./utils/paperColour";
@@ -37,7 +37,7 @@ export { getPageNumPaperOrientation, getPageNodePaperOrientation } from "./utils
 export { getPageNumPageMargins } from "./utils/pageRegion/margins";
 export { getPageNumPageBorders, getPageNodePageBorders } from "./utils/pageBorders";
 export { getBodyNodeMargins } from "./utils/nodes/body";
-export { doesDocHavePageNodes } from "./utils/page";
+export { doesDocHavePageNodes } from "./utils/nodes/page";
 
 // === Exports ===
 export { PageNode, HeaderFooterNode, BodyNode };
